+++ src/swap/defi/uni-v2-based/plugins/velodrome.ts
@@ -125,7 +125,6 @@ export function makeVelodromePlugin(
         tokenId: request.fromTokenId,
         spendTargets: [
           {
-            memo: swapTx.data,
             nativeAmount:
               swapTxs.length === 2 && i === 0
                 ? '0' // approval transactions don't have a value
@@ -133,6 +132,10 @@ export function makeVelodromePlugin(
             publicAddress: swapTx.to
           }
         ],
+        memos:
+          swapTx.data != null
+            ? [{ type: 'hex', value: swapTx.data.replace(/^0x/, '') }]
+            : undefined,
         customNetworkFee: {
           gasPrice:
             swapTx.gasPrice != null