+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - added: Login info to testconfig for testpartners.ts
 - changed: Change Lifi and Thorchain DA to use variable quotes
-- fixed: LiFi and Thorchain/DA memo handling
+- fixed: Memo handling by DEX plugins
 
 ## 2.3.1 (2024-03-29)