+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added: Login info to testconfig for testpartners.ts
 - changed: Change Lifi and Thorchain DA to use variable quotes
 - fixed: Memo handling by DEX plugins
+- fixed: Letsexchange orderUri
 
 ## 2.3.1 (2024-03-29)